requirements.txt
selenium>=4.1.0
jinja>=1.1
